Unset CROSS_COMPILE environment variable in build scripts
Tools makefiles recognize this variable, so we need to unset it, otherwise standard libraries will be be built with toolchain pointed by CROSS_COMPILE, which is undesirable. For example if it is set to little endian toolchain and script builds big endian toolchain then standard libraries will be little endian rendering toolchain unusable. In more dangerous case endianness will match, but standard libraries will be build with other compiler and there would be no easy way to notice this.