Avoid repeated reboots of nsim_gdb
nSIM outputs messages to std(out|err) as it accepts new connections. This
output is ignored is by this script, however it seems that buffer that contains
output at some points stops dropping old output to make a place for new one and
simply overflows.  Thus sim_gdb cannot write to stream, it is blocked and it
stops responding to TCP connection and Dejagnu is forced to restart it. I don't
know a way to redirect nsim_gdb output to null from Expect, so we have to use a
shell to help as here.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>
(cherry picked from commit f49ee2da2b68360828c5eb48f33dffa4764f8d90)