Do not do Git operations on directories from source tarball
Decide default value for auto-pull/checkout based on whether toolchain
directory is a Git repository or not. Checkout and pull if it is a
repository and do nothing if it is a directory from a source tarball.

Signed-off-by: Anton Kolesov <anton.kolesov@synopsys.com>
(cherry picked from commit bde7ed355eae2dbcdc4a3f34539d2289619b97ef)