Rename DejaGNU board arc-sim-mdb to arc-sim-nsimdrv
This commit renames DejaGNU board arc-sim-mdb to arc-sim-nsimdrv. Also this
script now will invoke nsimdrv directly instead of using MDB and wrapper
script. Wrapper script has been converting exit code of MDB to text message
"EXIT <exit_code>", which is useless now, because nsimdrv doesn't return
application exit code as it's own exit code. For this reason it status
wrapper is required for this board and it is being switched on in this
commit. As a part of this change, logic in this script has been changed to
conform to difference between MDB and nsimdrv.

Signed-off-by: Anton Kolesov <anton.kolesov@synopsys.com>