Add compatibility with Buildroot
Previously it was impossible to use our toolchain as an external toolchain
of Buildroot. Buildroot expects toolchain to support sysroot functionality,
which wasn't the case with our tools. It looks like that recently Buildroot
dropped this a requirements, however our toolchain still was incompatible
due to different directory structure.

This commit:
1. Adds --with-sysroot option to gcc configure script.
2. Installs uClibc headers and lib files to sysroot location.
3. Installs kernel headers to sysroot
4. Copies libstdc++ and libgcc to sysroot as well. It seems like it is not
really required and everything works anyway. But here I'm mimicking what
Buildroot does, and this might be useful in case we would want to have
toolchain with multiple sysroots.
5. HARDWIRED_ABSPATH is uClibc configuration is disabled as well to
simplify creation of relocatable toolchains.
6. As a results of this rel-rpaths.sh script will not attempt to patch
libc.so anymore, since that is not really required.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>