Fix problem with missing headers from uClibc
After the move to the sysroot-based toolchain, uClibc headers were absent
from `sysroot/usr/include`. Instead there were in the `sys-include`. It was
happening because stage 1 toolchain has been installed in the temporary
directory and thus stage 2 toolchain has been configured with option
`--with-headers`. This configure option means that those headers will be
copied to the sys-include. That however is not what we want - we want them
to be in the sysroot/usr/include. Several changes has been made:

1. Both stage 1 and stage 2 toolchains are installed at the same directory,
so stage 2 overwrites stage 1. This is the way of Buldroot and doesn't seem
to cause any problems (other than stage 1 is not available for archiving
purposes).

2. Stage 1 however is now built in the separate directory. I've observed
build errors when stage 2 was using the same build directory as stage 1.

3. Linux headers now are installed only ones. Previously that has been done
two times - first into the stage 1 `include` directory, then into the stage
2 `include`. Second installation has been causing issues, because Linux has
been removing headers files in the existing directories, even of those files
has been coming from uClibc.

4. uClibc is now configured with the proper RUNTIME_PREFIX and DEVEL_PREFIX.
Variable PREFIX is now used to ensure that uClibc is installed at proper
location. Value of RUNTIME_PREFIX is hardcoded into the ld-uClibc.so, so
previously we were hardcoding the path on the build host into the target
dynamic loader. Note that PREFIX is a variable in Makefile, but not in the
.config, so it is passed directly to make and is not written into .config.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>