Only support using defconfig files for uClibc configuration.
In the past a mechanism was in place to use a file uClibc/arc_config as
the configuration file for uClibc.  For over a year now uClibc has
instead been configured using defconfig files.

The defconfig mechanism takes care of lots of dependency issues that
using a fixed arc_config files does not handle (the process of
converting from defconfig to .config does dependency resolution of the
config settings) and so the arc_config mechanism should not be used any
more.

The patch removes support for arc_config.

ChangeLog:

	* build-uclibc.sh: Remove use of arc_config.