uclibc configuration: Create and use a temporary defconfig.
The process of creating a .config file from a defconfig file performs
dependency checks and resolution on the configuration options.  Editing
the .config file is discouraged as knowledge of the dependencies between
configuration options would need to be coded into the edit script, which
would be prone to breaking as uClibc dependencies changed over time.

A better approach is to make changes to the defconfig file and then use
the edited defconfig to build the .config in the normal way.

This patch switches to using a temporary defconfig file which is then
deleted once the .config file has been created.

Tested by comparing the generated .config files from before and after
this change.

ChangeLog:

	* build-uclibc.sh: Create a temporary defconfig file, edit the
	temporary file, then create the .config file.