Use separate toolchain configuration file
Use separate file to specify which branches (or tags) to checkout. This is
basic principle of information technology - keep data and algorithms
separate. Previous setup has been causing a situation where repository
contained multiple branches only for the sole purpose of providing custom
set of git branches. This is no more, as now it is possible to just add
configuration file and use one extra option of build-all.sh.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>