DejaGNU: Set correct password
There are issues with FTP procedures when running regressions, naimly an error
'can not find channel named "exp5"' in ftp_download. It seems that they have
been happening because ftdp expects password even when it is empty, but
DejaGNU has was sending it like `user $user $pass`, so when password is an
empty string, ftpd was treating like an absense of password in command,
since there are no quaotes. To workaround this issue I set `ftp_passwrod` to
\'\', so DejaGNU will send it like `user $user ''`. That will make sure that
ftpd will treat this as an empty assword.

It looks like that this problem poped up after rebasing to the latest
buildroot release.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>