Fix uClibc build
uClibc has switched to NPTL by default, which requires TLS, that is not
enabled by default in toolchain.This patch will make sure then when --tls is
not used, than NPTL will not be built. That is a temporary patch, as I think we
should replace --tls with --nptl to be more relevant, and we want to be ON
by default.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>