Always quit GDB after running a dejagnu test case
That a third take on solving an issue with targets not being reset properly.
It relies on GDB server (nSIM in this case) to implement k-packet to reset
target. Typically gdb-comm.exp board used for nSIM-GDB testing reuses GDB
instance, however that causes troubles in some cases:

1. Unfortunate combination of PC+BLINK registers causes a failed assertion
in generic GDB code. Target should be reset before GDB disconnects from it.

2. GDB sometimes fails to properly load debug symbols from new file, at
least when this file has asme name as before, but different contents. This
problem is sporadic, so I'm not sure what are exact conditions for
reprodcution, but having same time stamaps doesn't cause issue. To avoid
this GDB should be restarted completely between tests. While restarting it
will also send k-packet to target, which will rest it.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>