Fix libstdc++ build problem in Canadian cross
When building toolchain for Windows with Canadian cross, there is an error
"fenv.h not found" at stage of precompiled headers compilation. This happens
because in Canadian cross when gcc and libstdc++ are built separately from
newlib, libstdc++ is configured using the host (already built) compiler
using headers of this compiler, but when newlib is built as well - headers
of host compiler are replaced with newlib ones. But when doing header
precompilation host compiler is suddenly used without it's own headers, which
causes dissonance with configuration stage. I think that ultimately this is
a problem with top-level configure scripts, because path to newlib headers
is passed to it, and option --newlib is used, but it simply checks for
whether newlib is built by it or not, so it doesn't take into accound those
option here.

To workaround the problem we just disable precompiled header for Canadian
cross tools. Like it is done in crosstool-ng:

https://github.com/crosstool-ng/crosstool-ng/blob/master/config/cc/gcc.in.2#L173-182

Interestingly I see that precompiled headers are not installed anyway even
when built, I'm not sure maybe we use some invalid make target to install,
but I wasn't able to find what would install them. That kind of begs for
question why they are needed at all, as the only user of not-installed
headers would be libstdc++ regression test suite, when run from the build
tree. So it seems it would be fine to not build them in all cases, but to
avoid unexpected problem I disable them only for Canadian cross tools,
leaving rest in its current state.

Also see this GCC PR about this topic:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=40974

That however doesn't solve our problems, as it adds -nostdinc++ when
building headers, but not when configuring libstdc++.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>