Generate Linux installer with IDE + toolchain
OpenOCD is not part of this yet. First step would be to build OpenOCD from
the same Makefile.release as well, then to copy it into installer.

Unfortunately it seems to be not possible to install ARC plugin into Eclipse
for Win32 on Linux hosts - dependency on host system goes too deep and it is
not possible to run it at all. Eclipse help

http://help.eclipse.org/indigo/topic/org.eclipse.platform.doc.isv/guide/p2_director.html

suggests that it should be possible to do that using Eclipse for host, but
that didn't worked for me.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>