Merge NSIS files
We used to have two flavours of installers and for this purpose there were
several NSIS file: with common code and specific to each flavour. But since
now there is only one flavour of installer, there is no need to keep
separate files, especially since one of them was very small.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>