Build native gdb automatically
We always used to build gdbserver for ARC, but automatic building of full
GDB hasn't been implemented. This patch replaced old-schoold gdbserver build
with a new shiny process, that will build both GDB and gdbserver. GDB
requires ncurses - that is build automatically. Native GDB build can be disabled
via option --no-native-gdb. In the latter case GDBserver still will be built
- it doesn't need any external dependencies, so there should be no harm in
building it.

Note that we used to build gdbserver with static linkage, but now it is
linked dynamically, regardless of whether native GDB is built or not. Such
gdbserver will not work if there is an error in dynamic linker - gdbserver
will not start in this case. However such cases are very rare, therefore I
don't think it makes sense to sacrifice 99.9% of cases for this one rare
case.  If somebody needs static gdbserver - they can build it manually, it's
not a rocket science.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>