windows: Force user to uninstall previous IDE
It is known that sometimes installing new IDE on top of older installation
can break new installation. Eclipse is especially fragile. In addition
previous installation might include some custom files which might be
overwritten.

It is possible to start uninstaller from GUI of new installer, but we do not
do this here, mostly because it is expected that after uninstalling user
will have to uninstall leftover files from eclipse

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>