windows: Add Finish page to uninstaller
Back in a time a warning message has been added to uninstaller that was
saying that some files are removed by it - files that were created by
Eclipse, for example. However it seems that this message wasn't in the
version imported to GitHub, and nobody noticed for one and half year. This
commits returns this message. See STAR 9000691488.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>