Add option to build expat
Add option on whether to use expat library installed into system or build it
in this script. expat is used by GDB to parse XML, therefore is needed to
work with XML Target descriptions. Usually using system expat is sufficient,
however it may cause issues if system doesn't have expat or it is of wrong
version.  Notable use case for --no-system-expat is mingw32 build, if mingw
installation lacks expat. Default value is to use system expat.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>