Fix EM7D linker script compatibility with basic crtbegin.o
There was a linker failure if em7d.lds is used without custom crt0.S file:

    ld: Error: Overflow detected in relocation value;
    Relocation value should be between 16777215 and -16777216 whereas it
    268435644
    Local symbol: ".text".
    Relocation type is:R_ARC_S25W_PCREL
    FileName:crtbegin.o
    Section Name:.init

That was happening because .init section was in ICCM but .text was in RAM,
so the jump between them wasn't fitting into S25 in crtbegin.o. We cannot
move .init into RAM, as it is expected to be 0x0 aperture, so instead I move
.text into ICCM. This doesn't happen with a custom crt0.S file - there a
32-bit jump is used. Interestingly this doesn't happen with arcv2elfb
emulation and similar memory.x file - there error doesn't happen and I'm
not sure why.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>