Remove local temp repository from Eclipse p2 metadata
When ARC plugins are installed into eclipse with script, Eclipse adds
installation source location as a repository, even though nobody asks it to
make it permanent. As a result Eclipse for ARC is shipped with enable
repository that references location on my computer. Therefore "Check for
updates" command in Eclipse fails. Not nice.

I wasn't able to find how to remove repository from Eclipse with some
command line commands, so resorted to plain editing of files that contain
inappropriate information.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>