Build newlib and libstdc++ with optimization for size
Only for baremetal toolchain. That will configure newlib "nano" options and
will build it with compiler options that produce least text and data size.
Note that newlib-nano is not C-standard confirming, therefore some
particular features might not work, although newlib-nano developers believe
that those features are not useful in embedded system (at least when code
size is a concern).

Libstdc++ is built with -fno-exceptions. This follows ARM-guidelines for
newlib-nano.

libgcc is also built with -f{function,data}-section with this commit.
However there is no libgcc version optimized for size.

To use newlib nano add --specs=nano.specs to linker options. As a general
advice it is strongly recommended to link with --gc-sections options - both
newlib-nano and libstdc++ are built with -ffunction-sections
-fdata-sections.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>