Use nSIM hostlink from libgloss in DejaGNU boards
Support of nSIM hostlink was part of the newlib/libc. That is not quite
right, so hostlink related functions have been moved into libgloss. Now they
are in the separate library file, therefore are not linked in by default -
library should be specified separately. This is done via specs option of GCC
driver.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>