DejaGNU: Enable nSIM Turbo when available
Two libstdc++ numerics tests take quite a lot of time to execute on nSIM
without Turbo and without hardware floating point - it is around 15 minutes
on my machine with Xeon@3.5GHz. Enabling nSIM Turbo reduces time to a
passable 1 minute, however Turbo is an extra feature and nSIM exits with an
error when Turbo is in the option, but is not supported.

This patch introduces a check for Turbo availability: when it is present in
nSIM, then it is used. If Turbo is not supported, then timeout will be
increased to be big enough for those tests to pass.

This is applied only to libstdc++ testsuite, because this is the only test
suite that is known to benefit from Turbo. Other tests actually have some
slight regression in the run time. It is also undesirable to increase
timeout without a justified need, because that would be an issue if there
are lots of tests that are failing due to timeout, and which are caused by a
genuine error.

Other option would be to increase timeout only for those particular tests,
however that requires changes to the test case source file, and I'm not sure
that would be easy to get upstream (this can be done by adding a
dg-timeout-factor directive). Therefore I preferred to do change globally in
the board file.

libstdc++ tests that need this big timeout:

26_numerics/random/binomial_distribution/operators/values.cc
26_numerics/random/negative_binomial_distribution/operators/values.cc

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>