Mention spec files in documentation
With introduction of libgloss support it is now required to pass an
appropriate libgloss library that implements essential functions. In most cases
it is better to pass it via specific "specs" files since those will handle
all other things, including but not limited to need to wrap libc and
libgloss as a single group and/or properly support nano libraries.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>