Pass nsim_download_elf_sections only when needed
nSIM option "-on nsim_download_elf_sections" is needed to avoid overhead of
memory initialization of heap and stack sections. This option was introduced
in nSIM 2015.09, on older nSIM version it causes a warning message.
Unfortunately this warning causes a test failure at least in one case - in
"newlib.stdlib/atexit.c output". Hence this patch adds a check for nSIM
version, so this option will be passed only to versions that recognize it.

File revision.txt is not present in verification builds of nSIM, hence to
DejaGNU script has to check for it's existence before reading. If file
doesn't exist, then it is assumed that this is the "latest" version of nSIM
- which requires nsim_download_elf_sections option.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>