Use cleaner way to create list of nSIM options
There is a lot if string appending in DejaGNU boards to construct nSIM
options. This used to be done via a straight string appending of TCL, but
this patch replaces this with a list appending with further join. At least
now there is no need to repeat variable name two times to append an element.
Instead of this:

    string nsim_flags "$nsim_flag -on nsim_fast"

we now have this:

    lapend nsim_flags -on nsim_fast

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>