Use bash explicitly
Code that has been copied from ARM build scripts uses bashism which doesn't
work on Ubuntu. Therefore this patch changes used shell interpretor to
"bash" instead of "sh" - same as it is done in ARM toolchain.

The only script that needs this is build-elf32.sh, yet this change is
applied to all relevant scripts to preserve consistency. I didn't applied
this to script, that I don't think will be used by end users. However this
is not applied to arc-external.sh - this script requires "-e", so that
caller will be aware of download failure, but it seems that
`/usr/bin/env bash -e` doesn't work on RHELs.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>