Fix cross-compiling of libs optimized for size
When building libs optimized for size, target compiler is invoked to get a
list of multilibs. This invocation is not possible when cross-compiling. In
that case we invoke ARC compiler that runs on host - it is already in the
PATH because it is needed to build libraries.

This means that arc-elf32-gcc compiler in the PATH must have same multilibs
as the one that is being cross-compiled.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>