Support cross building for Linux/uClibc toolchain
This patch adds support of cross compilation to build-uclibc.sh. With this
patch build-uclibc.sh will respect --host option passed to build-all. Note
that it is required that ARC Linux/uClibc for the build host must be present
in the PATH to build standard C libraries for toolchain. Those changes had
to be done:

* Disable stage 1 build, instead ARC tools from the PATH are used to build
  libraries.
* Disable C++ header precompilation, similar to how it is done for
  baremetal toolchain.
* Set STRIPPROG explicitly when building binutils and GDB - both of them
  fail to set it properly at some places.
* Do not copy libgcc_s.so and libstdc++*.so libraries - this step is needed
  only for cross-compile toolchain, which has a sysroot.

This has been tested by building toolchain for ARC HS Linux with host ==
"arc-snps-linux-uclibc" - that is self hosting toolchain for ARC Linux.

Sysroot is not needed for self hosting toolchain.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>