Move crt, libgcc and libstdc++ to sysroot
As mentioned in code comments - some .o, .a and .so files are not installed
in sysroot as they should, even despite that --with-sysroot is used. Thus
files have to be moved manually into sysroot to produce a sysroot that is
completely self-contained and thus enable usage of multiple sysroot. Without
this patch even when --sysroot= options is passed to GCC it still would use
libgcc and others from it's own library locations - something that doesn't
work for us, where sysroots may contains different architectures. This is
changed by this patch, so not it is completely possible to use ARC HS
toolchain to produce ARC 700 applications using the sysroot from ARC 700
toolchain. And vice versa.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>