Release: Build dynamically linked toolchains
We used to build toolchain for releases with static linkage. This had was
having two issues:

    a) Less efficient memory usage due to lack of shared libraries;
    b) GCC wasn't supporting plugins.

This patch sets that toolchain for release should be linked dynamically. For
windows cross toolchain it also copies prerequisite DLL files to folders
that contain executable files. This is kept in Makefile.release instead of a
more generic build-elf32.sh because current implementation supports only a
specific Mingw toolchain, therefore might not work with other Mingw builds,
and or other target hosts (although this is barely needed for Linux hosts).
Possibly in the future that would be integrated into build-elf32.sh or
build-all.sh, as a more generic implementation. For now users building
dynamically linked toolchain in Canadian cross should copy necessary
libraries (if any) manually.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>