Windows: Fix shortcut link to documentation folder
Shortcut to documentation folder (share/doc) was created in "common" section
that is executed before "toolchain" sections that create this folder.
Therefore shortcut was pointing to location as it is a file, not a folder,
so shortcut wasn't working. To fix that shortcut should be created when
folder already exists.

This patch moves shortcut creation and removal into toolchain sections -
there are two of those for little and big endian. So if user installs both,
then shortcut will be "created" and removed two times, but NSIS should
handle this well.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>