Smartly detect for which ARC processor we simulate using the compiler macros.
This patch renders absolete the requirements of having the environment
variable ARC_MULTILIB_OPTIONS defined. Thus, when it is not defined,
the default procesor will be used. The default processor is given by
--with-cpu flag passed to the configure script when we build the
compiler.

The simulator parameters are set by procedures that are checking the
gcc default defines. Hence, regardless of the mcpu option, we will get
the correct ARC family. The mcpu option will be changed to hold
multiple cpu templates (e.g. mcpu=em4).

2016-01-12  Claudiu Zissulescu  <claziss@synopsys.com>

	* dejagnu/baseboards/arc-sim-nsimdrv.exp: Select simulation
	options based on the gcc macros.
	* dejagnu/tool-extra.exp: New file.