Fix build error for ARC HS Linux toolchain
There was a build error when building Linux toolchain for ARC HS, at the
stage of uClibc building:

    GEN include/bits/sysnum.h
    cc1: warning: multiple -mcpu= options specified. [enabled by default]
    cc1: warning: multiple -mcpu= options specified. [enabled by default]
    <stdin>:1:0: error: Illegal option for arc700 architecture
    <stdin>:1:0: error: Illegal option for arc700 architecture
    <stdin>:1:0: error: FPU options are available for ARC HS/EM only
    <stdin>:1:0: error: Illegal option for arc700 architecture
    <stdin>:1:0: error: Illegal option for arc700 architecture
    <stdin>:1:0: error: FPU options are available for ARC HS/EM only
    ERROR: Could not generate syscalls.

That was happening if value if --cpu option was not "archs", but something
else, like hs38_linux. In this case build-all.sh failed to identify that
target is an ARC HS and was using a uClibc defconfig for ARC 700, thus ARC HS
compiler was passed with options for ARC 700, which was causing those
errors. This fix modifies the check, so instead of checking if CPU is ARC HS
it checks if it is ARC 700 - for ARC 700 it is unlikely that anybody would
bother with custom -mcpu values, while for ARC HS it might make some sense,
therefore custom -mcpu values could be legitimate. Another way to decide
which defconfig to use, would be to move this check after gcc sources have
been checked out and to parse arc-cpus.def to see which --cpu values are for
ARC 700 and which arc for ARC HS. However this is slightly more complex and
not likely to be of any practical use, considering that ARC 700 is an old
processor.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>