Release: Create windows tarballs compatible with 7-zip
7-zip doesn't support hard links in tar files - they are turned into 0-byte
files, so if one untars windows tarballs with 7-zip - such toolchain simply
doesn't work. This is not a problem for release scripts, which rely on
Cygwin and proper tar. And those Windows tarballs are not released to
public, however they are available in internal repositories, and internal
users occasionally have issues with them due to 7-zip incompatibility.

FWIW, starting Cygwin just to untar file can be a too complex action to just
untar the file, so it is much better to produce windows tarballs that play
well with 7-zip.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>