Allow user to overwrite or not simulation environment context
This patch renders absolete the requirements of having the environment
variable ARC_MULTILIB_OPTIONS defined. Thus, when it is not defined,
the default procesor will be used. The default processor is given by
--with-cpu flag passed to the configure script when we build the
compiler. This is requiered when we want to run specific tests which are
checking for the default gcc driver parameters.

The simulator parameters can be addnotate from site.exp. Hence, we can ask
to have the simulation traces whenever we want.

2016-01-08  Claudiu Zissulescu  <claziss@synopsys.com>

	* dejagnu/baseboards/arc-sim-nsimdrv.exp (ARC_MULTILIB_OPTIONS):
	Now is optional.
	(nsim_flags): Can be overwrite from site.exp.