Release: Build OpenOCD for Windows in the main release file
Now that builds of OpenOCD for Windows can be done on same RHEL hosts that
do the build of reset of toolchain release components, there is no need for
a separate `Makefile.openocd`. Now build of OpenOCD for Windows is fully
integrated into the make `release.mk` and is done together with the rest of
the toolchain process.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>