Include full version id in the file names
We used to include only the release version in the distributable files, like
just 2016.03. So all release candidates had the same filenames. The
assumption is that final release candidate *is* the release, so we simply
use same ditributables for release. That, however was causing some issues:

* filenames were identical, so to maintain multiple versions, one had to
  store them in different directories. That wasn't a problem for me
  personally, because releases are stored each in their own directory
  anyway, but those, who need only a single tarball got unnecessary extra
  moves.
* It was problematic to include some definite build identifiers into the
  release, because it is not cool if final release would claim to be "-rc2".
  Initially I've added a generic "build number" into the version string,
  which didn't convey any info for the users, but we would be able to
  match build number with particular RC or engineering build. With this
  commit this goes away as we would have to do a separate build for final
  release anyway.

Previously we were reusing final RC for final release. This would be weird
with this patch, because all tarballs would include a "-rc" in their name.
We cannot simply rename tarballs, because that would break convention that
name of tarball matches name of top-level directory in it. Therefore we
would need to repackage tarballs with different directory and file names. At
this rate we are better of just redoing final release, instead of reusing
pieces of RC. This doesn't break the convention that there should be no
commits after final RC - so at the source level final RC and release should
match, it is just that binary stuff is rebuilt for release.

In the past release process required human actions, which tended to result
in the human mistakes which were not always discovered immediately, hence
there was a certain desire to avoid rebuilding of release distributables,
due to risk of introducing errors. Now that release process is fully
automated risk of process mistake is negligible, so we can do as many
release builds as needed.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>