Install binutils and gdb with explicit targets
It has appeared that just passing `install` is not good for gdb and
binutils. When option `--strip` is passed to the build-all.sh, then the
install target is named `install-strip` and that causes a build failure in
readline library inlucded in binutils and GCC. To avoid build error with
`--strip` we can't call generic `install-strip` on the binutils and gdb.

Another problem is that for GCC `install-strip` causes target libraries to
be stripped as well - this is undesired, because it is controlled by the
option `--elf32-strip-target-libs`. Therefore for GCC installation should be
handled by separate `install-host` and `install-target` targets.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>