Support uClibc-ng in the build scripts
This is a fix for STAR 9001009437 "Build scripts do not support uClibc-ng".

It wasn't possible to build uClibc-ng with previous version of build
scripts, because when installing libc headers, uClibc-ng invokes the
compiler to generate some headers, but uClibc hasn't been doing that. As a
result, current build process was failing, because libc headers were
installed before the stage 1 compiler has been built. However, it is not
possible to simply switch order of stage 1 and libc headers installation,
because libgcc depends on libc headers. So this patch splits stage 1 into
compiler build and libgcc build. This:

1. Linux headers installation
2. uClibc headers installation
3. Binutils
4. GCC and libgcc (stage 1)
5. uClibc build
6. GCC, libgcc and libstdc++ (stage 2)
7. GDB

is replaced with this:

1. Linux headers installation
2. Binutils
3. GCC (stage 1)
4. uClibc headers installation
5. libgcc (stage 1)
6. uClibc build
7. GCC, libgcc and libstdc++ (stage 2)
8. GDB

This page: http://preshing.com/20141119/how-to-build-a-gcc-cross-compiler/
proposes that it is not required build gcc and libgcc second time, instead
one can simply build only libstdc++ and other high-level libraries at that
stage. Notice, that this means using the build tree of the "stage 1"
compiler, because building "just" libstdc++ in the separate build tree still
will cause a build of the compiler in that tree, even if it will not be
installed. See, for example, how libstdc++ optimized for size is built by
build-elf32.sh. For the time being, I've decided to avoid experimenting with
this, due to time constraints, however we should consider this possibility
in the future - whether it works and doesn't break anything. For example in
the baremetal toolchain we had to add the second stage build in the past to
avoid some errors, but, perhaps, it is not really required to do a complete
rebuild there as well, and rebuild just some particular components would be
sufficient.

New order of build process doesn't break compatibility with classic uClibc.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>