Release: Omit -release suffix when creating final release tarballs
When release is created only the `arc-` prefix is stripped from the Git tag
name, so for tag arc-2016.09-eng005, filenames include 2016.09-eng005 in
them. This is not good for final releases, because Git tag is
arc-2016.09-release, so files will include 2016.09-release. This patch adds
a code that will remove `-release` from filenames part.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>