Build newlib pdf multiple times Texinfo >= 6.1
There is some issue with building newlib PDFs on Ubuntu 16.04. It might be
same as:

https://bugs.launchpad.net/ubuntu/+source/texlive-bin/+bug/1270396

PDF file is actually created, but exit code is 1, so it is not installed. As
a result we have to call make three times: one to build libc.pdf, one to
build libm.pdf and one to install them.

I assume that this is specific to Texinfo version 6.1 instead of being just
Ubuntu 16.04, so check is done against texinfo version.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>