Add -fomit-frame-pointer to GDB tests
GCC generates improper DWARF info when frame-pointer is involved - all
variables in debug information are specified relatively to FP register instead
of a virtual frame base, hence this information is invalid in prologue and
epilogue. Since GDB doesn't epilogue (at least as of now), some watchpoint
tests fail, because they get an extra watchpoint value change even though it is
not a variable that has changed, by a FP register in epilogue.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>