Fix build issue of native ncurses on Ubuntu
This is a continuation of commit:

    01fe2c8 Update to ncurses 6.0

which didn't fixed the original problem for Ubuntu hosts.

GCC 5 introduced some changes to the preprocessor output, that causes a
compilation error in files generated by ncurses. See GCC commit:

  https://gcc.gnu.org/ml/gcc-patches/2014-06/msg02185.html

To counter that, ncurses 6.0 added a check for GCC version, and now adds
an option -P if this compiler is used. However initial implementation of
this change was done incorrect - as a part of an algorithm that evaluates
a version number, there was usage of AWK with regular expression POSIX
classes, like [:space:]. Unfortunately, not all AWK implementations
support character classes; case in point is mawk, which is the default
AWK on Ubuntu, while gawk works just fine with those classes. Therefore,
on Ubuntu MKlib_gen.sh failed to properly recognize version of compiler
and kept generating files for an older GCC version. Ncurses already has
this fixed, but they didn't had a numbered release since then. There are
weekly official snapshoots at

  http://invisible-mirror.net/archives/ncurses/current/

but I don't feel sure about referencing it instead of the GNU servers. So
instead, we keep using official 6.0 release and patch it to have just
this one required change. Unfortunately, ncurses doesn't provide public
access to it's source control (if they use any), so I've just generated
diff patch for MKlib_gen.sh using original release and latest snapshot
(6.0-20160806).

This patch will not be needed after updating to ncurses 6.1, after it
will be released.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>